teste de integração dele e update com sucesso e falha no repository

+++ src/main/java/com/github/maikoncarlos/libraryapi/api/service/impl/BookServiceImpl.java
@@ -34,11 +34,17 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void delete(BookEntity book) {
+        if( book == null || book.getId() == null ){
+           throw new IllegalArgumentException(" Book id cant be null! ");
+        }
         repository.delete(book);
     }
 
     @Override
     public BookEntity update(BookEntity book) {
+        if( book == null || book.getId() == null ){
+            throw new IllegalArgumentException(" Book id cant be null! ");
+        }
         return repository.save(book);
     }
